functions/requirements.txt
firebase_functions~=0.1.0

# Core dependencies
firebase-admin>=6.2.0
google-cloud-storage>=2.10.0
google-generativeai>=0.3.0
pinecone-client>=2.2.4
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Development dependencies
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1

protoc_gen_openapiv2
pillow

# Hedera Specific
hedera-sdk-py==2.28.0

# Cryptography and Security
cryptography>=41.0.0

# Optional but Recommended
requests>=2.31.0
typing-extensions>=4.7.0

# For type hinting and development
mypy>=1.4.0
types-requests>=2.31.0

